+++ docs/api-reference/composite-layer.md
@@ -78,7 +78,7 @@ functions called.
 The default implementation returns `pickParams.info` without any change.
 
 
-##### `getForwardProps`
+##### `getBaseLayerProps`
 
 Returns an object with the key base `Layer` props that `CompositeLayer`s typically forward to their sublayers.
