+++ docs/get-started/getting-started.md
@@ -30,4 +30,5 @@ export MapboxAccessToken={Your Token Here} && npm start
 ## Remarks:
 
 * `luma.gl` will not be automatically installed with deck.gl. The reason is that an application must only include one copy of luma.gl. Please explicitly install `luma.gl` with your app (`npm install luma.gl`). This is similar to React components which typically only have "peer dependencies" on React, and it is the application's responsibility to actually select and install a specific React version.
-* `MapboxAccessToken` is the Mapbox token. For most examples shipped with deck.gl, geospatial data is rendered on top of Mapbox-powered basemap. Mapbox requires users to provide an access token before they serve map tiles to the user's browser. For more information about map token read the sections about [Using with Mapbox GL](/docs/get-started/using-with-mapbox-gl.md).
+* `MapboxAccessToken` is the Mapbox token. For most examples shipped with deck.gl, geospatial data is rendered on top of a Mapbox-powered basemap. Mapbox requires users to provide an access token before they serve map tiles to the user's browser. For more information about map token read the sections about [Using with Mapbox GL](/docs/get-started/using-with-mapbox-gl.md).
+* Running the examples against the local deck.gl code. The examples contain an `npm run start-local` script that allow you to build and run the examples against the local code in the repo (e.g. master, instead of an installed version). There are some caveats with this mode, for instance you must make sure to install in the deck.gl root folder before running `npm run start-local` in an example.
